designmate/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docopt==0.6.2
fastapi==0.135.3
folium==0.20.0
fonttools==4.61.1
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.14
langchain-core==1.2.24
langchain-openai==1.1.12
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.16.0
numpy==2.4.1
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.0
pipwin==0.5.2
plotly==6.5.2
protobuf==6.33.5
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyjsparser==2.7.1
pyogrio==0.12.1
pyparsing==3.3.2
PyPrind==2.11.3
pyproj==3.7.2
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
statsmodels==0.14.6
streamlit==1.54.0
streamlit-folium==0.26.1
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchdog==6.0.0
xgboost==3.1.3
xxhash==3.6.0
xyzservices==2025.11.0
zstandard==0.25.0
